[781] Rearrange 1st pillar of transaction engine - raw to hashed to canonical
