--- app/jobs/hashed_transaction_job/imports.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module HashedTransactionJob
-  class Imports < ApplicationJob
-    def perform
-      ::HashedTransactionService::RawPlaidTransaction::Import.new.run
-      ::HashedTransactionService::RawEmburseTransaction::Import.new.run
-    end
-  end
-end
