+++ app/jobs/transaction_engine_job/nightly.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module TransactionEngineJob
+  class Nightly < ApplicationJob
+    def perform
+      ::TransactionEngine::Nightly.new.run
+    end
+  end
+end
