--- app/services/raw_plaid_transaction_service/likely/disbursement.rb
@@ -1,9 +1,0 @@
-module RawPlaidTransactionService
-  module Likely
-    class Disbursement
-      def run
-        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hcb disburse%'")
-      end
-    end
-  end
-end
