--- app/services/raw_plaid_transaction_service/likely/donation.rb
@@ -1,9 +1,0 @@
-module RawPlaidTransactionService
-  module Likely
-    class Donation
-      def run
-        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hackc donate%' or plaid_transaction->>'name' ilike '%hack club event%'")
-      end
-    end
-  end
-end
