--- app/services/raw_plaid_transaction_service/likely/emburse.rb
@@ -1,9 +1,0 @@
-module RawPlaidTransactionService
-  module Likely
-    class Emburse
-      def run
-        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%emburse%' or plaid_transaction->>'merchant' ilike '%emburse%'")
-      end
-    end
-  end
-end
