--- app/services/raw_plaid_transaction_service/likely/invoice.rb
@@ -1,9 +1,0 @@
-module RawPlaidTransactionService
-  module Likely
-    class Invoice
-      def run
-        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hackc payout%' or plaid_transaction->>'name' ilike '%hack club event%'")
-      end
-    end
-  end
-end
