+++ app/services/transaction_engine/canonical_transaction_service/map_to_event/process.rb
@@ -1,0 +1,25 @@
+module TransactionEngine
+  module CanonicalTransactionService
+    module MapToEvent
+      class Process
+        def run
+          likely_githubs.find_each do |ct|
+
+            ::TransactionEngine::CanonicalTransactionService::MapToEvent::Github.new(canonical_transaction: ct).run
+
+          end
+        end
+
+        private
+
+        def likely_githubs
+          ::CanonicalTransaction.exclude(excluded_ids).likely_github
+        end
+
+        def excluded_ids
+          @excluded_ids ||= ::CanonicalEventMapping.pluck(:canonical_transaction_id)
+        end
+      end
+    end
+  end
+end
