+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/ach.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Ach
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%busbillpay%'")
+        end
+      end
+    end
+  end
+end
