+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/check.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Check
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%dda#%' or plaid_transaction->>'name' ilike '%check%'")
+        end
+      end
+    end
+  end
+end
