+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/disbursement.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Disbursement
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hcb disburse%'")
+        end
+      end
+    end
+  end
+end
