+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/donation.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Donation
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hackc donate%' or plaid_transaction->>'name' ilike '%hack club event%'")
+        end
+      end
+    end
+  end
+end
