+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/emburse.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Emburse
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%emburse%' or plaid_transaction->>'merchant' ilike '%emburse%'")
+        end
+      end
+    end
+  end
+end
