+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/expensify.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Expensify
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%expensify%'")
+        end
+      end
+    end
+  end
+end
