+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/fee_payment.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class FeePayment
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%bank fee%'")
+        end
+      end
+    end
+  end
+end
