+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/fee_reimbursement.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class FeeReimbursement
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%fee refund%'")
+        end
+      end
+    end
+  end
+end
