+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/github.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Github
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%github grant%'")
+        end
+      end
+    end
+  end
+end
