+++ app/services/transaction_engine/raw_plaid_transaction_service/likely/invoice.rb
@@ -1,0 +1,11 @@
+module TransactionEngine
+  module RawPlaidTransactionService
+    module Likely
+      class Invoice
+        def run
+          ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%hackc payout%' or plaid_transaction->>'name' ilike '%hack club event%'")
+        end
+      end
+    end
+  end
+end
