+++ config/schedule.yml
@@ -3,17 +3,7 @@ g_suite_mark_verifieds_job:
   class: "GSuiteJob::MarkVerifieds"
   queue: "default"
   
-raw_plaid_transaction_imports_job:
+transaction_engine_nightly_job:
   cron: "0 6 * * *" # run at 6am UTC each morning
-  class: "RawPlaidTransactionJob::Imports"
-  queue: "default"
-
-raw_emburse_transaction_imports_job:
-  cron: "0 7 * * *" # run at 7am UTC each morning
-  class: "RawEmburseTransactionJob::Imports"
-  queue: "default"
-
-hashed_transaction_imports_job:
-  cron: "0 8 * * *" # run after the raw import jobs
-  class: "HashedTransactionJob::Imports"
+  class: "TransactionEngineJob::Nightly"
   queue: "default"
