+++ spec/fixtures/raw_plaid_transactions.yml
@@ -1,0 +1,7 @@
+raw_plaid_transaction1:
+  plaid_account_id: "plaid_account_id1"
+  plaid_item_id: "plaid_item_id1"
+  plaid_transaction_id: "plaid_transaction_id1"
+  amount_cents: 100
+  date_posted: "2020-09-02"
+  created_at: <%= Time.now %>
